+++ evident/__init__.py
@@ -1,4 +1,4 @@
-from .diversity_handler import UnivariateDataHandler, BivariateDataHandler
+from .data_handler import UnivariateDataHandler, BivariateDataHandler
 
 
 __version__ = "0.3.0"
